add 1-list_len.c

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -7,7 +7,7 @@
  */
 int list_len(const list_t *h)
 {
-    size_t count = 0;
+    int count = 0;
 
     while (h != NULL)
     {
